python/requirements.txt
openfeature-sdk==0.8.0
